+++ package.json
@@ -64,7 +64,7 @@
     ],
     "moduleNameMapper": {
       "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
-      "\\.(css|less|sass|scss)$": "<rootDir>/__mocks__/styleMock.js",
+      "\\.(css|less|sass|scss)$": "identity-obj-proxy",
       "^/(.*)$": "<rootDir>/src/$1",
       "^d3$": "<rootDir>/node_modules/d3/",
       "^react$": "<rootDir>/node_modules/react/",
@@ -191,6 +191,7 @@
     "html-loader": "1.3.2",
     "html-webpack-plugin": "4.5.1",
     "husky": "5.1.3",
+    "identity-obj-proxy": "3.0.0",
     "jest": "26.6.3",
     "jest-coverage-ratchet": "0.2.3",
     "jest-css-modules-transform": "4.2.0",
