+++ src/shared/components/results/styles/render-columns-in-card.module.scss
