+++ src/uniprotkb/components/protein-data-views/ProteinOverviewView.tsx
@@ -8,7 +8,7 @@ import { UniProtkbAPIModel } from '../../adapters/uniProtkbConverter';
 import OrganismDataView from '../../../shared/components/views/OrganismDataView';
 
 const ProteinOverview: FC<{
-  // Note: it would be good to eventually use RenderColumnInCard here
+  // Note: it would be good to eventually use RenderColumnsInCard here
   // which would involve either converting UniProtkbAPIModel to UniProtkbUIModel
   // or refactoring UniProtKBColumnConfiguration to use UniProtkbAPIModel.
   data: UniProtkbAPIModel;
